refactor(core-run): split profiling into a single file
